requirements.txt
# Core dependencies for automated blog generation
openai>=1.83.0
aiohttp>=3.12.0
praw>=7.8.0
pytrends>=4.9.0
beautifulsoup4>=4.13.0
lxml>=5.4.0
requests>=2.32.0
python-dotenv>=1.1.0
PyGithub>=2.6.0

# Testing dependencies
pytest>=8.4.0
pytest-asyncio>=1.0.0
pytest-cov>=6.1.0

# Data processing
pandas>=2.2.0
numpy>=2.2.0

# Existing Jekyll dependencies
Jinja2>=3.1.2
Markdown>=3.5.1
weasyprint>=60.1